get rid of more -Wconversion warnings

+++ buffer/buffer_feed.c
@@ -1,14 +1,14 @@
 #include "buffer.h"
 
-extern int buffer_stubborn_read(ssize_t (*op)(),int fd,const char* buf, size_t len,void* cookie);
+extern ssize_t buffer_stubborn_read(ssize_t (*op)(),int fd,const char* buf, size_t len,void* cookie);
 
-int buffer_feed(buffer* b) {
+ssize_t buffer_feed(buffer* b) {
   if (b->p==b->n) {
-    int w;
+    ssize_t w;
     if ((w=buffer_stubborn_read(b->op,b->fd,b->x,b->a,b))<0)
       return -1;
-    b->n=w;
+    b->n=(size_t)w;
     b->p=0;
   }
-  return (b->n-b->p);
+  return (ssize_t)(b->n-b->p);
 }
