+++ buffer/buffer_get_token_pred.c
@@ -7,7 +7,7 @@ ssize_t buffer_get_token_pred(buffer* b,char* x,size_t len,
   unsigned int blen;
 
   for (blen=0;blen<len;++blen) {
-    register int r;
+    register ssize_t r;
     if ((r=buffer_getc(b,x))<0) return r;
     if (r==0) break;
     if (p(x-blen,blen+1)) break;
