+++ buffer/buffer_getc.c
@@ -1,9 +1,9 @@
 #include "byte.h"
 #include "buffer.h"
 
-int buffer_getc(buffer* b,char* x) {
+ssize_t buffer_getc(buffer* b,char* x) {
   if (b->p==b->n) {
-    register int blen;
+    register ssize_t blen;
     if ((blen=buffer_feed(b))<=0) return blen;
   }
   *x=b->x[b->p];
