+++ open/readclose.c
@@ -2,19 +2,19 @@
 #include <errno.h>
 #include "readclose.h"
 
-int readclose_append(int fd,stralloc *sa,size_t bufsize)
+ssize_t readclose_append(int fd,stralloc *sa,size_t bufsize)
 {
-  int r;
+  ssize_t r;
   for (;;) {
     if (!stralloc_readyplus(sa,bufsize)) { close(fd); return -1; }
     r = read(fd,sa->s + sa->len,bufsize);
     if (r == -1) if (errno == EINTR) continue;
     if (r <= 0) { close(fd); return r; }
-    sa->len += r;
+    sa->len += (size_t)r;
   }
 }
 
-int readclose(int fd,stralloc *sa,size_t bufsize)
+ssize_t readclose(int fd,stralloc *sa,size_t bufsize)
 {
   if (!stralloc_copys(sa,"")) { close(fd); return -1; }
   return readclose_append(fd,sa,bufsize);
