+++ socket/fmt_ip6if.c
@@ -3,8 +3,8 @@
 #include "fmt.h"
 #include "socket.h"
 
-unsigned int fmt_ip6if(char* dest,const char* ip,uint32 scope_id) {
-  int i=fmt_ip6(dest,ip);
+size_t fmt_ip6if(char* dest,const char* ip,uint32 scope_id) {
+  size_t i=fmt_ip6(dest,ip);
   if (scope_id) {
     if (dest) {
       dest[i]='%'; ++i; dest+=i;
