+++ taia/taia_half.c
@@ -4,8 +4,8 @@
 
 void taia_half(tai6464* t,const tai6464* u) {
   t->atto = u->atto >> 1;
-  if (u->nano & 1) t->atto += 500000000UL;
+  if (u->nano & 1) t->atto += (uint32)500000000UL;
   t->nano = u->nano >> 1;
-  if (u->sec.x & 1) t->nano += 500000000UL;
+  if (u->sec.x & 1) t->nano += (uint32)500000000UL;
   t->sec.x = u->sec.x >> 1;
 }
