+++ taia/taia_now.c
@@ -23,7 +23,7 @@ void taia_now(struct taia *t)
   struct timeval now;
   gettimeofday(&now,(struct timezone *) 0);
   tai_unix(&t->sec,now.tv_sec);
-  t->nano = 1000 * now.tv_usec + 500;
+  t->nano = (uint32)(1000 * now.tv_usec + 500);
   t->atto = 0;
 #endif
 }
